mergecode-agent/requirements.txt
langchain-core>=0.1.0
langgraph>=0.0.15
python-dotenv>=1.0.0
requests>=2.28.0
